Adds modules
